requirements.txt
rich>=13.7.0
PyYAML>=6.0.1
pyinstaller>=6.3.0
Pillow>=10.0.0